+++ models/contacts.js
@@ -2,6 +2,7 @@ const fs = require('fs/promises')
 const path = require('path');
 const shortid = require('shortid');
 
+
 const contactsPath = path.join(__dirname, "/contacts.json");
 
 const listContacts = async () => {
